refactor: validate PartySizeMin and PartySizeMax after spawning
Instead of putting it in canSpawn and checking it on every spawn cycle. That had the problem that units with specified PartySizeMin did not get considered for spawning during cycles which happened before that size was achieved, leading to an under-representation of such units. The new system allows them to spawn from the very beginning, and only checks for validity of their PartySizeMin and PartySizeMax after a complete spawn process. Then it removes spawnables that violated these conditions from the spawn process and then does a new spawn process with the remaining resources.