feat: add separate functions to validate current and future RatioMax
Adjust the logic of isWithinRatioMax and getSpawnWeight accordingly. The main change is that we add +1 to the referenced total as well as to the predicted total of the spawnable and base our calculations off of that. The negative weight case in the getSpawnWeight now occurs more frequently with this setup, so the log pertaining to that has been removed.