feat!: remove IsUsingTopPartyResources from def and set it automatically
The purpose of this field is to ensure that subparties (defined inside StaticSpawnables) generate their spawns based on their own resources instead of considering the top party's resources as their own pool when spawning/upgrading. Therefore, if the `spawn` function is called we know that this party needs to do a full spawn process and is a static or standalone subparty. If it were a party inside DynamicSpawnables then it would be approached via the spawn process of the parent party and won't enter this function.