feat: rewrite unit upgrade weight to be cleaner
This fixes an issue where units were adjusting their upgrade weight based on their position in the parent's DynamicSpawnables array without regard to whether the parent is a UnitBlock or not as the this position only matters as the unit's tier when inside a UnitBlock.

While the new calculation will not produce identical results, it still follows the same philosophy as before i.e. UpgradeWeight is more if:
- Lower Cost
- More instances of the unit have spawned
- Lower tier